Renaming RequestPassWordForm.jsx to RequestPasswordForm.jsx
